+++ packages/network-msw/test/acceptance/acceptance.test.ts
@@ -1,0 +1,5 @@
+import { browserAcceptanceTests } from './browser-tests';
+import { resolve } from 'path';
+
+const testPackageRoot = resolve(__dirname, './dummy-react-app');
+describe('acceptance', browserAcceptanceTests(testPackageRoot, '3333'));
