+++ packages/network-msw/test/acceptance/dummy-react-app/public/index.html
