+++ packages/network-msw/test/acceptance/dummy-react-app/public/mockServiceWorker.js
