+++ packages/network-msw/test/acceptance/dummy-react-app/src/App.js
