+++ packages/network-msw/test/acceptance/dummy-react-app/src/index.js
