+++ packages/network-msw/test/acceptance/dummy-react-app/src/mocks/browser-handler.js
@@ -1,0 +1,1 @@
+../../../browser-handler.js
