+++ packages/network-msw/test/acceptance/dummy-react-app/src/mocks/browser.js
@@ -1,0 +1,5 @@
+import { setupWorker, rest } from 'msw';
+import { mswResolver } from '@graphql-mocks/network-msw';
+import { graphqlHandler } from './browser-handler';
+
+export const worker = setupWorker(rest.post('/graphql', mswResolver(graphqlHandler)));
