+++ packages/network-pretender/test/acceptance/acceptance.test.ts
@@ -1,0 +1,5 @@
+import { browserAcceptanceTests } from '../../../network-msw/test/acceptance/browser-tests';
+import { resolve } from 'path';
+
+const packageRoot = resolve(__dirname, './dummy-react-app');
+describe('acceptance', browserAcceptanceTests(packageRoot, '3232'));
