+++ packages/network-pretender/test/acceptance/dummy-react-app/package.json
@@ -8,11 +8,11 @@
     "react-scripts": "4.0.3"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "SKIP_PREFLIGHT_CHECK=true react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "serve": "http-server build -p 5555"
+    "serve": "http-server build -p 3232"
   },
   "browserslist": {
     "production": [
@@ -25,9 +25,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "msw": {
-    "workerDirectory": "public"
   },
   "devDependencies": {
     "http-server": "^13.0.1"
