+++ packages/network-pretender/test/acceptance/dummy-react-app/public/index.html
