+++ packages/network-pretender/test/acceptance/dummy-react-app/src/App.js
