+++ packages/network-pretender/test/acceptance/dummy-react-app/src/browser-handler.js
@@ -1,0 +1,1 @@
+../../../../../network-msw/test/acceptance/browser-handler.js
