+++ packages/network-pretender/test/acceptance/dummy-react-app/src/index.js
