--- packages/network-pretender/test/unit/index.test.ts
@@ -1,7 +1,0 @@
-import { expect } from 'chai';
-
-describe('index', function () {
-  it('works', async function () {
-    expect(true).to.be.true;
-  });
-});
